app/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.3.0
backports.tarfile==1.2.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
crashtest==0.4.1
cryptography==44.0.0
databases==0.9.0
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
dulwich==0.22.7
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.7
fastapi-jwt-auth==0.5.0
fastjsonschema==2.21.1
filelock==3.16.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.5
jiter==0.8.2
keyring==25.6.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
openai==1.59.6
orjson==3.10.14
packaging==24.2
passlib==1.7.4
pkginfo==1.12.0
platformdirs==4.3.6
poetry==2.0.0
poetry-core==2.0.0
propcache==0.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic-extra-types==2.10.1
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==1.7.1
pyproject_hooks==1.2.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
RapidFuzz==3.11.0
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9
SecretStorage==3.3.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
trove-classifiers==2025.1.10.15
typer==0.15.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.28.1
watchfiles==1.0.4
websockets==14.1
yarl==1.18.3
zipp==3.21.0
